--- dataSources/RelationAbstract/src/main/java/cn/edu/tsinghua/iginx/RelationAbstract/tools/IDataTypeTransformer.java
@@ -1,10 +1,0 @@
-package cn.edu.tsinghua.iginx.RelationAbstract.tools;
-
-import cn.edu.tsinghua.iginx.thrift.DataType;
-
-public interface IDataTypeTransformer {
-
-  public DataType fromEngineType(String dataType);
-
-  public String toEngineType(DataType dataType);
-}
