+++ dataSources/pom.xml
@@ -34,9 +34,8 @@
         <module>mongodb</module>
         <module>opentsdb</module>
         <module>parquet</module>
-        <module>postgresql</module>
         <module>redis</module>
-        <module>RelationAbstract</module>
+        <module>relational</module>
     </modules>
 
     <properties>
