+++ dataSources/relational/src/main/java/cn/edu/tsinghua/iginx/relational/exception/RelationalTaskExecuteFailureException.java
@@ -1,0 +1,14 @@
+package cn.edu.tsinghua.iginx.relational.exception;
+
+import cn.edu.tsinghua.iginx.engine.physical.exception.PhysicalTaskExecuteFailureException;
+
+public class RelationalTaskExecuteFailureException extends PhysicalTaskExecuteFailureException {
+
+  public RelationalTaskExecuteFailureException(String message) {
+    super(message);
+  }
+
+  public RelationalTaskExecuteFailureException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}
