+++ dataSources/relational/src/main/java/cn/edu/tsinghua/iginx/relational/tools/IDataTypeTransformer.java
@@ -1,0 +1,10 @@
+package cn.edu.tsinghua.iginx.relational.tools;
+
+import cn.edu.tsinghua.iginx.thrift.DataType;
+
+public interface IDataTypeTransformer {
+
+  public DataType fromEngineType(String dataType);
+
+  public String toEngineType(DataType dataType);
+}
