+++ test/pom.xml
@@ -46,7 +46,7 @@
         </dependency>
         <dependency>
             <groupId>cn.edu.tsinghua</groupId>
-            <artifactId>postgresql</artifactId>
+            <artifactId>relational</artifactId>
         </dependency>
         <dependency>
             <groupId>cn.edu.tsinghua</groupId>
